# perf: batch up series instead of always series of 1 in rowwise udf