## Changes Made

previously, `call_func_with_evaluated_exprs` was returning a PySeries with a single value. This was problematic because we were doing `Series::concat` on `N` series with a single value, and Series::concat is quite expensive. 

Instead, I refactored it so that the `call_func_with_evaluated_exprs` returns the python value, then we use the same `PySeries::from_pylist` method to create a series of many values. We're still performing a `Series::concat` at the end, but of much less values. 

here's a small test I ran that tokenizes the yelp_review_full dataset. 


```py
# %%
import daft
import time
import pyarrow as pa
from transformers import AutoTokenizer
from datasets import load_dataset

tokenizer = AutoTokenizer.from_pretrained("sentence-transformers/all-MiniLM-L6-v2")

@daft.func
def tokenize(s: str) -> list[int]:
    return tokenizer.encode(s)

# using datasets api because of hf caching bug
ds = load_dataset("Yelp/yelp_review_full")
tbl = pa.Table.from_batches(ds.data['train']._batches)
df = daft.from_arrow(tbl).collect()

start = time.perf_counter()

df = df.select(tokenize(df["text"])).collect()

end = time.perf_counter()
print(f"Execution time: {end - start:.4f} seconds")
```

### Results
```
# HEAD Execution time: 35.9118 seconds
# HEAD Execution time: 35.6711 seconds
# HEAD Execution time: 36.0107 seconds

# main Execution time: 49.6138 seconds
# main Execution time: 48.4861 seconds
```

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to `docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130 for docs review)