@universalmind303 2025-08-06
I added a few additional optimizations in https://github.com/Eventual-Inc/Daft/pull/4920/commits/251ea588fc6ba2fc0a704c18ef77fd1bf62cc282

namely: 

- removing `Series::concat` entirely
a few micro optimizations 
- reuse a vec per chunk instead of creating it in the hot loop
- remove a few additional allocations 


The original tests are down to ~34 seconds now 
```
# Execution time: 33.6569 seconds
# Execution time: 33.9771 seconds
# Execution time: 33.3724 seconds
```